requirements.txt
pytest==7.3.0
coverage==7.2.3
PyYAML==6.0
pyinstaller==5.10.0
python-magic==0.4.27
termcolor==2.3.0
python-pam==2.0.2
six==1.16.0
pycryptodome==3.19.0